requirements.txt
aiohttp~=3.7.4.post0
PyYAML~=5.4.1
pydantic~=1.8.2